+++ changelog.md
@@ -8,6 +8,7 @@
 
 ### Other New Features
 - weather service pirateweather.net as replacement for darksky.net (thanks to aschwith) 
+- improved unit handling in weather services via language files
 
 ### Improvements
 - Template checker now ignores multi-line comments  
@@ -33,6 +34,7 @@
 ### Fixed Bugs
 - multiple icons of icon.heating on one page could only be colored with the color of the first occurance and disappeared otherwise
 - colorizing in basic.print did not change colors after first change (due to faulty removeClass() statement)
+- wind speed in openweathermap service was in m/s but unit was km/h
  
 ### Known Bugs
 - if item contains a stringified number (e.g. with leading zero). widget.set converts it back to numeric format - so basic.print can not print it as text